fix: unify buttons, refactor PlaceBet component
